+++ src/MicroElements.Functional/SharedUtils/TypeExtensions.cs
@@ -2,22 +2,47 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Reflection;
 
 namespace MicroElements.Functional
 {
+    /// <summary>
+    /// Reflection extensions.
+    /// </summary>
     public static class TypeExtensions
     {
+        /// <summary>
+        /// Gets default value for type.
+        /// </summary>
+        /// <param name="type">Source type.</param>
+        /// <returns>Default value.</returns>
         public static object? GetDefaultValue(this Type type)
         {
             return type.IsValueType ? Activator.CreateInstance(type) : null;
         }
 
-        public static object? GetDefaultValueCompiled(this Type type)
+        /// <summary>
+        /// Determines whether <paramref name="type"/> is assignable to <typeparamref name="T" />.
+        /// </summary>
+        /// <typeparam name="T">The type to test assignability to.</typeparam>
+        /// <param name="type">The type to check.</param>
+        /// <returns>True if type is assignable to references of type <typeparamref name="T" />; otherwise, False.</returns>
+        public static bool IsAssignableTo<T>(this Type type)
         {
-            Func<Unit, object> func = CodeCompiler.CachedCompiledFunc<Unit, object>(type, GetDefaultValueInternal<CodeCompiler.GenericType>);
-            return func(Unit.Default);
+            type.AssertArgumentNotNull(nameof(type));
+
+            return typeof(T).GetTypeInfo().IsAssignableFrom(type.GetTypeInfo());
         }
 
-        internal static object GetDefaultValueInternal<T>(Unit unit) => default(T);
+        /// <summary>
+        /// Determines whether <paramref name="type"/> is concrete type: (not interface and not abstract).
+        /// </summary>
+        /// <param name="type">The type to check.</param>
+        /// <returns>True if type is concrete.</returns>
+        public static bool IsConcreteType(this Type type)
+        {
+            type.AssertArgumentNotNull(nameof(type));
+            return !type.IsInterface && !type.IsAbstract;
+        }
     }
 }